add image upload and delete feature

+++ src/restaurants/restaurants.controller.ts
@@ -1,9 +1,11 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, UsePipes, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Put, Query, UploadedFiles, UseInterceptors } from '@nestjs/common';
+import { FilesInterceptor } from '@nestjs/platform-express';
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
 import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurant.schema';
+import { Multer } from 'multer';
 
 @Controller('restaurants')
 export class RestaurantsController {
@@ -34,8 +36,23 @@ export class RestaurantsController {
 
     @Delete(':id')
     async deleteRestaurant(@Param('id') id: string): Promise<{ deleted: Boolean }> {
-        const restaurant = this.restaurantsServices.deleteById(id)
-        if (!restaurant) return { deleted: true }
+        const restaurant = await this.restaurantsServices.findById(id)
+        const isDeleted = await this.restaurantsServices.deleteImages(restaurant.images)
+        if (isDeleted) {
+            await this.restaurantsServices.deleteById(id)
+            return { deleted: true }
+        }
         else return { deleted: false }
+    }
+
+    @Put('upload/:id')
+    @UseInterceptors(FilesInterceptor('files'))
+    async uploadFile(
+        @Param('id') id: string,
+        @UploadedFiles() files: Array<Express.Multer.File>
+    ) {
+        await this.restaurantsServices.findById(id)
+        const res = await this.restaurantsServices.uploadImages(id, files)
+        return res
     }
 }
